requirements.txt
python-dotenv==1.0.0
openai==1.3.5
beautifulsoup4==4.12.2
python-dateutil==2.8.2
email-validator==2.0.0.post2
pyfiglet==0.8.post1
colorama==0.4.6
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.10.1
joblib==1.3.2
google-api-python-client==2.108.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
